오케스트레이션 메소드 자동화
