+++ src/main/java/com/example/demo/mapper/GameDataMapper.java
@@ -20,7 +20,4 @@ public interface GameDataMapper {
 	public void margeNintendoDC();
 	public void margeNintendoNew();
 	public void margePlanetNew();
-	
-	public void scrapMargeClean();
-	public void scrapMarge();
 }
