+++ src/main/java/com/example/demo/mapper/GameInfoMapper.java
@@ -1,7 +1,5 @@
 package com.example.demo.mapper;
 
-import java.util.List;
-
 import org.apache.ibatis.annotations.Mapper;
 
 import com.example.demo.model.GameInfo;
@@ -12,8 +10,5 @@ public interface GameInfoMapper {
 	public void scrapMargeClean();
 	public void scrapMarge(GameInfo gameinfo);
 
-	public void gameInfoClean();
 	public void updateGameInfo();
-	public List<GameInfo> gameInfoList();
-
 }
